devtools/server/tests/chrome/test_css-properties.html

<!DOCTYPE HTML>
<html>
<!--
Bug 1265798 - Replace inIDOMUtils.cssPropertyIsShorthand
-->
<head>
  <meta charset="utf-8">
  <title>Test CSS Properties Actor</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  const { initCssProperties } = require("devtools/client/fronts/css-properties");

  function toSortedString(array) {
    return JSON.stringify(array.sort());
  }

  // Return some mocked values for a tab. Override hasActor with the
  // options passed into the test. A second property is used
  // during these tests: `win` for css-properties.js's getClientBrowserVersion.
  async function getTabProperties(url, useActor) {
    const { target, win } = await attachURL(url);
    target.hasActor = () => useActor;
    return { target, win };
  }

  const runCssPropertiesTests = async function(url, useActor) {
    info(`Opening two tabs ${useActor ? "with" : "without"} CssPropertiesActor support.`);

    // Open a new tab. The only property we are interested in is `target`.
    const tab = await getTabProperties(url, useActor);
    const { cssProperties } = await initCssProperties(tab);

    ok(cssProperties.isKnown("border"),
      "The `border` shorthand property is known.");
    ok(cssProperties.isKnown("display"),
      "The `display` property is known.");
    ok(!cssProperties.isKnown("foobar"),
      "A fake property is not known.");
    ok(cssProperties.isKnown("--foobar"),
      "A CSS variable properly evaluates.");
    ok(cssProperties.isKnown("--foob\\{ar"),
      "A CSS variable with escaped character properly evaluates.");
    ok(cssProperties.isKnown("--fübar"),
      "A CSS variable unicode properly evaluates.");
    ok(!cssProperties.isKnown("--foo bar"),
      "A CSS variable with spaces fails");

    is(toSortedString(cssProperties.getValues("margin")),
       toSortedString(["auto", "inherit", "initial", "unset", "revert"]),
       "Can get values for the CSS margin.");
    is(cssProperties.getValues("foobar").length, 0,
      "Unknown values return an empty array.");

    const bgColorValues = cssProperties.getValues("background-color");
    ok(bgColorValues.includes("blanchedalmond"),
      "A property with color values includes blanchedalmond.");
    ok(bgColorValues.includes("papayawhip"),
      "A property with color values includes papayawhip.");
    ok(bgColorValues.includes("rgb"),
      "A property with color values includes non-colors.");
  };

  addAsyncTest(async function setup() {
    const url = document.getElementById("cssProperties").href;
    await runCssPropertiesTests(url, true);
    await runCssPropertiesTests(url, false);

    runNextTest();
  });

  SimpleTest.waitForExplicitFinish();
  runNextTest();
};
  </script>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1265798">Mozilla Bug 1265798</a>
  <a id="cssProperties" target="_blank" href="inspector_css-properties.html">Test Document</a>
</body>
</html>
